requirements.txt
bleak>=0.22
numpy>=1.26
rumps>=0.4.0
pyobjc-core>=10.3; platform_system == "Darwin"
pyobjc-framework-Cocoa>=10.3; platform_system == "Darwin"
pyobjc-framework-Quartz>=10.3; platform_system == "Darwin"
